requirements.txt
django==4.2.19
django-environ==0.12.0
psycopg2-binary==2.9.10
# psycopg2==2.9.5
pillow==11.1.0
django-filter==24.3
# Fabric3@git+https://github.com/ahmedyasserays/fabric.git@ff01d6b7c65ca6477261a92765e25b5f36fa8dfa
gunicorn==23.0.0
djangorestframework==3.15.2
# django-allauth==64.0.0
djangorestframework-simplejwt==5.4.0
# dj-rest-auth==6.0.0
requests==2.32.3
drf-yasg==1.21.8
django-cors-headers==4.7.0
black==24.10.0
django-ckeditor==6.7.2
pytest==8.3.5
pytest-django==4.11.1
django-prometheus==2.3.1
Faker==18.3.0